refactor: give each rule a dedicated file
