+++ src/rules/padding-around-before-all-blocks.ts
@@ -1,0 +1,16 @@
+import { PaddingType, StatementType, createPaddingRule } from './utils/padding';
+
+export const config = [
+  {
+    paddingType: PaddingType.Always,
+    prevStatementType: StatementType.Any,
+    nextStatementType: StatementType.BeforeAllToken,
+  },
+  {
+    paddingType: PaddingType.Always,
+    prevStatementType: StatementType.BeforeAllToken,
+    nextStatementType: StatementType.Any,
+  },
+];
+
+export default createPaddingRule(config);
