+++ src/rules/padding-around-before-each-blocks.ts
@@ -1,0 +1,16 @@
+import { PaddingType, StatementType, createPaddingRule } from './utils/padding';
+
+export const config = [
+  {
+    paddingType: PaddingType.Always,
+    prevStatementType: StatementType.Any,
+    nextStatementType: StatementType.BeforeEachToken,
+  },
+  {
+    paddingType: PaddingType.Always,
+    prevStatementType: StatementType.BeforeEachToken,
+    nextStatementType: StatementType.Any,
+  },
+];
+
+export default createPaddingRule(config);
