+++ src/rules/padding-around-expect-groups.ts
@@ -1,0 +1,21 @@
+import { PaddingType, StatementType, createPaddingRule } from './utils/padding';
+
+export const config = [
+  {
+    paddingType: PaddingType.Always,
+    prevStatementType: StatementType.Any,
+    nextStatementType: StatementType.ExpectToken,
+  },
+  {
+    paddingType: PaddingType.Always,
+    prevStatementType: StatementType.ExpectToken,
+    nextStatementType: StatementType.Any,
+  },
+  {
+    paddingType: PaddingType.Any,
+    prevStatementType: StatementType.ExpectToken,
+    nextStatementType: StatementType.ExpectToken,
+  },
+];
+
+export default createPaddingRule(config);
