+++ src/rules/padding-around-test-blocks.ts
@@ -1,0 +1,28 @@
+import { PaddingType, StatementType, createPaddingRule } from './utils/padding';
+
+export const config = [
+  {
+    paddingType: PaddingType.Always,
+    prevStatementType: StatementType.Any,
+    nextStatementType: [
+      StatementType.TestToken,
+      StatementType.ItToken,
+      StatementType.FitToken,
+      StatementType.XitToken,
+      StatementType.XtestToken,
+    ],
+  },
+  {
+    paddingType: PaddingType.Always,
+    prevStatementType: [
+      StatementType.TestToken,
+      StatementType.ItToken,
+      StatementType.FitToken,
+      StatementType.XitToken,
+      StatementType.XtestToken,
+    ],
+    nextStatementType: StatementType.Any,
+  },
+];
+
+export default createPaddingRule(config);
